+++ src/live.cpp
@@ -84,6 +84,12 @@ LiveBuilder& LiveBuilder::SetCompression(Compression compression) {
   return *this;
 }
 
+LiveBuilder& LiveBuilder::SetSlowReaderBehavior(
+    SlowReaderBehavior slow_reader_behavior) {
+  slow_reader_behavior_ = slow_reader_behavior;
+  return *this;
+}
+
 databento::LiveBlocking LiveBuilder::BuildBlocking() {
   Validate();
   if (gateway_.empty()) {
@@ -91,14 +97,14 @@ databento::LiveBlocking LiveBuilder::BuildBlocking() {
                                    dataset_,        send_ts_out_,
                                    upgrade_policy_, heartbeat_interval_,
                                    buffer_size_,    user_agent_ext_,
-                                   compression_};
+                                   compression_,    slow_reader_behavior_};
   }
   return databento::LiveBlocking{log_receiver_,   key_,
                                  dataset_,        gateway_,
                                  port_,           send_ts_out_,
                                  upgrade_policy_, heartbeat_interval_,
                                  buffer_size_,    user_agent_ext_,
-                                 compression_};
+                                 compression_,    slow_reader_behavior_};
 }
 
 databento::LiveThreaded LiveBuilder::BuildThreaded() {
@@ -108,14 +114,14 @@ databento::LiveThreaded LiveBuilder::BuildThreaded() {
                                    dataset_,        send_ts_out_,
                                    upgrade_policy_, heartbeat_interval_,
                                    buffer_size_,    user_agent_ext_,
-                                   compression_};
+                                   compression_,    slow_reader_behavior_};
   }
   return databento::LiveThreaded{log_receiver_,   key_,
                                  dataset_,        gateway_,
                                  port_,           send_ts_out_,
                                  upgrade_policy_, heartbeat_interval_,
                                  buffer_size_,    user_agent_ext_,
-                                 compression_};
+                                 compression_,    slow_reader_behavior_};
 }
 
 void LiveBuilder::Validate() {